refactor(round-3): extract detectLanguage from db.ts to dedicated module #92
The 60-line filename → language map was sitting at the bottom of the
1780-line db.ts and had been flagged as a refactor candidate in
CLAUDE.md. Move it to src/server/detect-language.ts. Documents in the
new module why it stays separate from src/languages.ts (different
target vocabulary: storage labels vs. PrismJS grammar keys).

Pure code move + import — no behavior change.
